Feature API
This commit adds an endpoint for fetching features for a given client.
The client is obtained from the x-api-key header value that is passed
on each request, since it is unique across all clients, it can give
access without an elaborated authentication system.

The endpoint works on url /api/features

Using a new module named Api assigned to Feature api endpoints
seems to resolve the issue where the view with name was not found.
Maybe it was a naming strategy related issue.